requirements.txt
# Install these requirements if you wish to contribute to this project.

# ===========================================================================
# Packages essential to this app. Needed by anyone who wants to use this app.
# ===========================================================================
django==1.4.3
django-cms==2.3.5
django-filer==0.9.3
cmsplugin-filer==0.9.3
simple-translation==0.8.6
South==0.7.6
django-hvad==0.3


# ==============================================================
# Packages needed for running the tests. Needed by contributors.
# ==============================================================
Fabric
django-libs
factory-boy
mock
django-nose
coverage
django-coverage
ipython
ipdb
